@@ -5,27 +5,27 @@
   "dependencies": {
     "classnames": "^2.2.6",
     "lodash": "^4.17.15",
-    "monday-sdk-js": "^0.1.2",
-    "node-sass": "^4.13.1",
+    "monday-sdk-js": "^0.1.4",
     "react": "^16.10.2",
     "react-dom": "^16.10.2",
     "react-modal": "^3.11.2",
     "react-onclickoutside": "^6.9.0",
-    "react-scripts": "3.2.0",
+    "react-scripts": "^5.0.1",
     "request": "^2.88.2",
+    "sass": "^1.54.8",
     "typeface-roboto": "^0.0.75",
-    "webpack": "4.41.0"
+    "webpack": "^5.0.0"
   },
   "devDependencies": {
+    "@mondaydotcomorg/tunnel": "^0.4.1",
     "concurrently": "^4.1.2",
-    "cross-port-killer": "^1.2.1",
-    "ngrok": "^3.2.7"
+    "cross-port-killer": "^1.2.1"
   },
   "scripts": {
     "start": "npm run stop && concurrently \"npm run server\" \"npm run expose\"",
     "server": "react-scripts start",
     "build": "react-scripts build",
-    "expose": "ngrok http 8304",
+    "expose": "mtunnel -p 8304",
     "stop": "kill-port 8304 && kill-port 4049 && kill-port 4040"
   },
   "eslintConfig": {
