@@ -1,36 +1,43 @@
 import React from "react";
+import { useState, useEffect } from "react";
 import "./App.css";
 import mondaySdk from "monday-sdk-js";
-import "monday-ui-react-core/dist/main.css"
+import "monday-ui-react-core/dist/main.css";
 //Explore more Monday React Components here: https://style.monday.com/
-import AttentionBox from "monday-ui-react-core/dist/AttentionBox.js"
+import AttentionBox from "monday-ui-react-core/dist/AttentionBox.js";
 
+// Usage of mondaySDK example, for more information visit here: https://developer.monday.com/apps/docs/introduction-to-the-sdk/
 const monday = mondaySdk();
 
-class App extends React.Component {
-  constructor(props) {
-    super(props);
+const App = () => {
+  const [context, setContext] = useState();
 
-    // Default state
-    this.state = {
-      settings: {},
-      name: "",
-    };
-  }
+  useEffect(() => {
+    // Notice this method notifies the monday platform that user gains a first value in an app.
+    // Read more about it here: https://developer.monday.com/apps/docs/mondayexecute#value-created-for-user/
+    monday.execute("valueCreatedForUser");
 
-  componentDidMount() {
-    // TODO: set up event listeners
-  }
+    // TODO: set up event listeners, Here`s an example, read more here: https://developer.monday.com/apps/docs/mondaylisten/
+    monday.listen("context", (res) => {
+      setContext(res.data);
+    });
+  }, []);
 
-  render() {
-    return <div className="App">
+  //Some example what you can do with context, read more here: https://developer.monday.com/apps/docs/mondayget#requesting-context-and-settings-data
+  const attentionBoxText = `Hello, your user_id is: ${
+    context ? context.user.id : "still loading"
+  }.
+  Let's start building your amazing app, which will change the world!`;
+
+  return (
+    <div className="App">
       <AttentionBox
         title="Hello Monday Apps!"
-        text="Let's start building your amazing app, which will change the world!"
+        text={attentionBoxText}
         type="success"
       />
-    </div>;
-  }
-}
+    </div>
+  );
+};
 
 export default App;
