@@ -1,10 +1,9 @@
-const localtunnel = require('localtunnel');
+const mondayTunnel = require('@mondaydotcomorg/tunnel').default;
 
 const MAX_RETRIES = 5;
 
 const createTunnel = async (port, retries = 0) => {
-  const tunnel = await localtunnel({
-    host: process.env.TUNNEL_SERVER_HOST,
+  const tunnel = await mondayTunnel({
     port,
     subdomain: process.env.TUNNEL_SUBDOMAIN,
   });
@@ -23,10 +22,6 @@ const createTunnel = async (port, retries = 0) => {
     console.warn('could not use requested subdomain, generated a random one');
   }
   console.log(`listening at localhost:${port} || tunnel: ${url}`);
-
-  tunnel.on('close', () => {
-    // tunnels are closed
-  });
 };
 
 module.exports = { createTunnel };
